[python] Applies relocation principle which implies to have empty __init__.py Modifies meta_taks_6d and meta_tasks_kine accordingly.
