--- src/dynamic_graph/sot/core/__init__.py
@@ -1,21 +1,0 @@
-# flake8: noqa
-from .exp_moving_avg import *
-from .feature_generic import *
-from .feature_joint_limits import *
-from .feature_point6d import *
-from .feature_pose import *
-from .feature_position import *
-from .feature_position_relative import *
-from .feature_posture import *
-from .feature_visual_point import *
-from .gain_adaptive import *
-from .gradient_ascent import *
-from .joint_limitator import *
-from .kalman import *
-from .math_small_entities import *
-from .op_point_modifier import *
-from .robot_simu import *
-from .sot import *
-from .task import *
-from .task_pd import *
-from .visual_point_projecter import *
