+++ src/dynamic_graph/sot/core/meta_tasks_kine.py
@@ -4,7 +4,7 @@ from dynamic_graph.sot.core.meta_task_6d import MetaTask6d
 # TODO: this function is imported from meta_tasks_kine in several places, whereas it is defined in meta_tasks
 from dynamic_graph.sot.core.meta_tasks import gotoNd  # noqa
 from dynamic_graph.sot.core.meta_tasks import MetaTaskCom
-from dynamic_graph.sot.core.sot import Task
+from dynamic_graph.sot.core.task import Task
 
 
 class MetaTaskKine6d(MetaTask6d):
